Moved ap10k dataset and formatting
